DTSRD-2504: add repository integration test
